Split backend into respective providers
